Fixed the bug of negative amount in Payment method

+++ src/ankhmorpork/Game/Game.java
@@ -1264,7 +1264,7 @@ Integer cardIndexNumber = PresentationUtility.returnRandomNumber(0,lstCardID.len
 			Player objPayToPlayer = GetPlayer(PayToPlayerID);
 			Player objPaidByPlayer =GetPlayer(PaidByPlayer);
 						
-			if(objPaidByPlayer.getPlayerAmount()>= Amount)
+			if((int)objPaidByPlayer.getPlayerAmount()>= Amount)
 			{
 				objPaidByPlayer.setPlayer_amount(objPaidByPlayer.getPlayer_amount()-Amount);
 				objPayToPlayer.setPlayer_amount(objPayToPlayer.getPlayer_amount()+ Amount);
@@ -1309,7 +1309,7 @@ Integer cardIndexNumber = PresentationUtility.returnRandomNumber(0,lstCardID.len
 		{			
 			Player objPaidByPlayer =GetPlayer(PaidByPlayer);
 						
-			if(objPaidByPlayer.getPlayerAmount()>= Amount)
+			if((int)objPaidByPlayer.getPlayerAmount()>= Amount)
 			{
 				objPaidByPlayer.setPlayer_amount(objPaidByPlayer.getPlayer_amount()-Amount);
 				Game.GameBank.setBankAmount(Game.GameBank.getBankAmount()+ Amount);
@@ -1337,7 +1337,7 @@ Integer cardIndexNumber = PresentationUtility.returnRandomNumber(0,lstCardID.len
 				{			
 					Player objPayToPlayer =GetPlayer(PayToPlayer);
 								
-					if(Game.GameBank.getBankAmount()>= Amount)
+					if((int)Game.GameBank.getBankAmount()>= Amount)
 					{
 						objPayToPlayer.setPlayer_amount(objPayToPlayer.getPlayer_amount()+Amount);
 						Game.GameBank.setBankAmount(Game.GameBank.getBankAmount()- Amount);
